requirements.txt
asgiref==3.8.1
Django==5.0.4
django-crum==0.7.9
django-widget-tweaks==1.5.0
mysqlclient==2.2.4
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
